restapi: move Config to main.go
